o improve 'sessions'-command

+++ src/henplus/view/Column.java
@@ -1,7 +1,7 @@
 /*
  * This is free software, licensed under the Gnu Public License (GPL)
  * get a copy from <http://www.gnu.org/licenses/gpl.html>
- * $Id: Column.java,v 1.2 2004-01-27 18:16:34 hzeller Exp $ 
+ * $Id: Column.java,v 1.3 2004-03-06 15:37:18 hzeller Exp $ 
  * author: Henner Zeller <H.Zeller@acm.org>
  */
 package henplus.view;
@@ -20,7 +20,7 @@ public class Column {
     private int width;
     private int pos;
 
-    public Column(int value) {
+    public Column(long value) {
         this(String.valueOf(value));
     }
